WIP getting metadata display and fix for loading metadata

+++ caper/caper/views.py
@@ -462,6 +462,7 @@ def profile(request, message_to_user=None):
 
     for proj in projects:
         prepare_project_linkid(proj)
+        proj['sample_metadata_available'] = has_sample_metadata(proj['_id'])
 
     prefs = get_user_preferences(request.user)
     form = UserPreferencesForm(prefs)
@@ -1807,8 +1808,10 @@ def edit_project_page(request, project_name):
             ## get extra metadata from csv first (if exists in old project), add it to the new proj
             new_id = _create_project(form, request, extra_metadata_file_fp, previous_versions = new_prev_versions, previous_views = [views, downloads])
 
-            if os.path.exists(temp_directory):
+            # can't delete it, if its being used in the other thread for metadata extraction
+            if os.path.exists(temp_directory) and not extra_metadata_file_fp:
                 shutil.rmtree(temp_directory)
+            
             if new_id is not None:
                 # go to the new project
                 return redirect('project_page', project_name=new_id.inserted_id)
@@ -2142,6 +2145,7 @@ def admin_stats(request):
 
     # Calculate stats
     for project in public_projects:
+        project['sample_metadata_available'] = has_sample_metadata(project['_id'])
         if 'project_downloads' in project:
             # Process download stats
             pass
@@ -2276,6 +2280,10 @@ def extract_project_files(tarfile, file_location, project_data_path, project_id,
         query = {'_id': ObjectId(project_id)}
         if extra_metadata_filepath:
             runs = process_metadata_no_request(replace_underscore_keys(runs), file_path=extra_metadata_filepath)
+            parent_dir = os.path.dirname(extra_metadata_filepath)
+
+            if os.path.exists(parent_dir):
+                shutil.rmtree(parent_dir)
 
         new_val = {"$set": {'runs': runs,
                             'Oncogenes': get_project_oncogenes(runs)}}
